Minor fixes after Eclipse integration
